Move selectors to selectors.ts
