+++ enterprise/frontend/src/embedding-sdk/hooks/public/use-question-search.ts
@@ -1,7 +1,7 @@
 import { useMemo } from "react";
 
 import { useSdkSelector } from "embedding-sdk/store";
-import { getIsInitialized, getIsLoggedIn } from "embedding-sdk/store/reducer";
+import { getIsInitialized, getIsLoggedIn } from "embedding-sdk/store/selectors";
 import { useSearchListQuery } from "metabase/common/hooks";
 
 export const useQuestionSearch = (searchQuery?: string) => {
